# handler/org/out.go
package org

type CreateResponse struct {
	IsCreated bool `json:"is_created"`
} // @name CreateResponse

type UpdateResponse struct {
	IsUpdated bool `json:"is_updated"`
} // @name UpdateResponse

type DeleteResponse struct {
	IsDeleted bool `json:"is_deleted"`
} // @name DeleteResponse